requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

python-multipart>=0.0.9,<1.0.0


# Gradio frontend
gradio==4.44.1


# CrewAI and agents
crewai==0.28.8
crewai-tools==0.1.6



# File processing
python-docx==1.1.0
PyPDF2==3.0.1

# HTTP client for Gradio
requests==2.31.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0
